idmap in place for retailers and dc (as they are mutable), id map for tranactions deleted as they are immutable, CocaColaHQ is now tied to the framework and is in the db, transactions are lazy loaded

+++ src/main/java/mappers/IdentityMap.java
@@ -23,10 +23,8 @@ public class IdentityMap<E> {
         map.put(id, obj);
     }
 
-    public E get(long id) {
+    public E get(int id) {
         return map.get(id);
     }
-
-
 
 }
